app/requirements.txt
bcrypt
boto3
fastapi
load_dotenv
psycopg2-binary
pydantic
python-dotenv
python-multipart
sqlalchemy
uvicorn[standard]